Add a script to remove pem files from source archive
This change adds a shell script that removes the pem files that are used
to generate KEK and db authentication keys. This is for legal reasons as
we are not supposed to distribute them.

This script is intended to be used before pushing any new source archive
of varstored to our git repository. That way, both RPM and SRPM files
built by our build server won't contain the pem files and can be
distributed safely.

The script takes no argument and gets the tarball file name from the
Source0 entry in the spec file. It then uncompresses it, deletes the pem
files from the uncompressed tarball, and recompresses it once done.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>