Check for pem files in %prep stage
This change searches for pem files in the build folder and stops the
build if any is found. Pem files present means they also are in the
source archive and must be removed with the remove-certs-from-tarball.sh
script located in the SOURCES folder. This will prevent us from
accidentally distribute RPM and SRPM with pem files.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>